requirements.txt
python-docx 
networkx 
openai
pillow
matplotlib
requests
backoff
cohere
tiktoken
nltk
numpy